Add files via upload
This is the dataset with all information from all sources. Many of the columns are only for commentary purposes. Sources are CMSWS (output-related variables), Mecklenburg OpenMapping (soil type and hydraulic groups - they are related), and USGS (monthly average precipitation between 2010 and 2020 in inches). Note that the dataset is completely raw so you have to calculate pipe age in python by either using the construction date or the inventory date. 
If you have any further questions, email jq1@uncc.edu.